requirements.txt
langchain
pinecone-client
streamlit
python-dotenv
openai
tiktoken